Fix deleteLastDigit and isStringANumber
deleteLastDigit was converting non-number values to '0', this is wrong.
If the calculator is not showing a number, i.e. an error message like
'overflow', then this method should not alter that value.

deleteLastDigit now calls isStringANumber to determine if the
displayValue is indeed a valid number.

isStringANumber required a minor fix to allow values with a trailing
decimal point like '5.', this is a valid value to show on the
calculator.

Non-number tests were updated for deleteLastDigit and 1 test was added
to isStringANumber to cover the trailing decimal case.