Change default values set by resetCalculator
It makes more sense to set the operand and operator variables to null
here, since this now provides a real initial state for the calculator.
The calculator behaves a little differently in the initial state, this
makes it easier to deal with how that is handled in other methods.

The resetDisplayValue should be false here since it is a brand new value
waiting for the user to enter something. The resetDisplayValue being
true would mean the current display value is a result value and if the
user enters a new digit, then that display value will be cleared. That
seems unnecessary at the initial state.