Add resetDisplayValue check to inputNegativeSign
When the resetDisplayValue is set to true, it means that the value
currently set to displayValue (the number that is actually shown on the
calculator) is currently showing a result number instead of an input
number.  The difference between a result number and an input number is
that when the user enters a new number, the following would happen in
each state:

  result number:  leftOperand = rightOperand
                  rightOperand = result number on the screen.
                  Screen gets cleared and new number is entered.
                  resetDisplayValue = false

  input number:  New number digit gets appended to the number on the
                 screen, unless the number of digits is currently at max
                 count, then it is ignored.

When resetDisplayValue is true (result mode), it is expected that the
leftOperand and displayValue are the same.  The reason is the way the
algorithm works, when a calculation is made, it will take the
leftOperand and rightOperand and execute whatever the current operator
is.  That new value then gets assigned to be leftOperand and
displayValue; leftOperand is an integer while displayValue is a string.
The reason for this way is that if the user now starts repeatedly
hitting the equal key, it will just perform the same operation using the
last number entered.  For example 3 + 2 will be 5 after hitting equal
key, but if the equal key continues to get hit, the values would
increment to 7, 9, 11, etc.

Back to the inputNegativeSign method, which is what gets called when the
user selects the "+/-" key (flips the number's sign between positive and
negative).  If the display is currently in "result number" mode, the
sign change needs to carry over to the leftOperand as well.

If the current displayValue is 0 while in "result number" mode, then
nothing happens since you cannot have a -0, so there is no test to check
for the leftOperand.  Note that there are tests that already exist when
the displayValue is 0 and inputNegativeSign is called, but those are
just checking to make sure the displayValue is correct and not looking
at the leftOperand.