Update the order of steps in inputOperator method
The assignment of the new operator now happens after the inputEquals
call.  Also will no longer set resetDisplayValue to true here.

Here is the reason for this change, if the user entered a sequence
like this on the calculator:

  1 + 2 -

Upon hitting the '-' key, the previous entries should resolve to 3.
Before this change was made, that would have resolved to -1.

Also, there is no reason to set resetDisplayValue here, the inputEquals
method will handle that.

Updated the unit tests that are related to this method as well. Those
will no longer check for the resetDisplayValue, since inputOperator no
longer sets this.