Fix issue with inputOperator not setting a value
The inputOperator was not setting resetDisplayValue to true.  When user
enters a number, then hits an operator, like '+', the input part is over
so resetDisplayValue needs to be true at this point.  Any further number
inputs will wipe what is showing on the display and start a new number.

Additional checks were added to the Unit Tests as well.